% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gvar.R
\name{compare_gvar}
\alias{compare_gvar}
\title{Compare Variance Components between Two Models}
\usage{
compare_gvar(
  fit_a,
  fit_b,
  cutoff = 5,
  dec_rule = "OR",
  n_draws = 1000,
  comp = "frob",
  return_all = FALSE
)
}
\arguments{
\item{fit_a}{Fitted model object for Model A}

\item{fit_b}{Fitted model object for Model B}

\item{cutoff}{The percentage level of the test (default: 5\%)}

\item{dec_rule}{The decision rule to be used. Currently only supports default "OR".}

\item{n_draws}{The number of draws to use for reference distributions (default: 1000)}

\item{comp}{The distance metric to use. Should be one of "frob" (Frobenius norm), "maxdiff" (maximum  difference), or "l1" (L1 norm) (default: "frob")}

\item{return_all}{Logical indicating whether to return all distributions (default: FALSE)}
}
\value{
A list containing the results of the comparison. The list includes:
 \itemize{
  \item{sig_beta}{The decision on whether there is a significant difference between the variance components for Model A and Model B (based on the beta parameter)}
  \item{sig_pcor}{The decision on whether there is a significant difference between the variance components for Model A and Model B (based on the partial correlation parameter)}
  \item{res_beta}{The null distribution for the variance components (based on the beta parameter) for both models}
  \item{res_pcor}{The null distribution for the variance components (based on the partial correlation parameter) for both models}
  \item{emp_beta}{The empirical distance between the two models (based on the beta parameter)}
  \item{emp_pcor}{The empirical distance between the two models (based on the partial correlation parameter)}
  \item{larger_beta}{The number of times the null hypothesis (based on the beta parameter) was rejected across all draws}
  \item{larger_pcor}{The number of times the null hypothesis (based on the partial correlation parameter) was rejected across all draws}
   }
}
\description{
Computes the empirical distance between two models based on the variance components
and compares them using reference distributions. Returns the p-value for the comparison
based on a decision rule specified by the user.
}
