% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{draws_array2matrix}
\alias{draws_array2matrix}
\title{Convert Draws Array to Matrix}
\usage{
draws_array2matrix(array_3d, warmup = 0)
}
\arguments{
\item{array_3d}{A 3D array of draws where each slice represents an iteration.}

\item{warmup}{An integer specifying the number of initial samples to discard as warm-up. Default is 0.}
}
\value{
A matrix where each row represents an iteration.
}
\description{
This helper function transforms a 3D array of draws into a matrix. 
It also allows for the removal of warmup samples.
}
