% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{draws_matrix2array}
\alias{draws_matrix2array}
\title{Convert Draws Matrix to Array}
\usage{
draws_matrix2array(draws_matrix)
}
\arguments{
\item{draws_matrix}{A matrix of draws where each row represents an iteration.}
}
\value{
A 3D array where each slice represents an iteration.
}
\description{
This helper function transforms a matrix of draws into a 3D array.
}
