% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{draws_matrix2list}
\alias{draws_matrix2list}
\title{Convert Draws Matrix to List of Matrices}
\usage{
draws_matrix2list(draws_matrix)
}
\arguments{
\item{draws_matrix}{A matrix of draws where each row represents an iteration.}
}
\value{
A list of matrices, each representing a single iteration.
}
\description{
This helper function transforms a matrix of draws into a list of matrices, 
each representing a single iteration.
}
