% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_gvar.R
\name{ess_gvar}
\alias{ess_gvar}
\title{Compute Effective Sample Sizes for MCMC Samples of BGGM GVAR models}
\usage{
ess_gvar(fitobj, burnin = 50)
}
\arguments{
\item{fitobj}{A list containing a BGGM fit object.}

\item{burnin}{An integer indicating the number of burn-in iterations to discard. Default is 50.}
}
\value{
A list with two elements: ess_beta and ess_pcor. ess_beta contains the ESS of MCMC samples of VAR, and ess_pcor contains the ESS of MCMC samples of partial correlation coefficients.
}
\description{
This function computes the effective sample sizes (ESS) of MCMC samples of VAR and partial correlation coefficients (pcor) based on the provided MCMC fit object.
}
