% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{get_centrality}
\alias{get_centrality}
\title{Compute Centrality Measures}
\usage{
get_centrality(fitobj, burnin = 500, remove_ar = TRUE)
}
\arguments{
\item{fitobj}{A fit object containing the beta and pcor samples.}

\item{burnin}{An integer specifying the number of initial samples to discard as burn-in. Default is 500.}

\item{remove_ar}{A logical value specifying whether to remove the autoregressive effects for centrality calculation. Default is TRUE.}
}
\value{
A list containing the following centrality measures:
\itemize{
  \item \code{instrength}: In-strength centrality.
  \item \code{outstrength}: Out-strength centrality.
  \item \code{strength}: Contemporaneous strength centrality.
  \item \code{density_beta}: Temporal network density.
  \item \code{density_pcor}: Contemporaneous network density.
}
}
\description{
This function computes various centrality measures for a given fit object.
}
\examples{
\dontrun{
fit <- # fit your model here TODO add this
  centrality_measures <- get_centrality(fit, burnin = 500)
}

}
