% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{plot_centrality}
\alias{plot_centrality}
\title{Plot Centrality Measures}
\usage{
plot_centrality(
  obj,
  plot_type = "tiefighter",
  density_type = "outstrength",
  cis = 0.95
)
}
\arguments{
\item{obj}{An object containing the centrality measures obtained from [get_centrality()].}

\item{plot_type}{A character string specifying the type of plot. Accepts "tiefighter" or "density". Default is "tiefighter".}

\item{density_type}{A character string specifying the type of density. Accepts "instrength", "outstrength", "strength", "density_beta", "density_pcor". Default is "outstrength".}

\item{cis}{A numeric value specifying the credible interval. Must be between 0 and 1 (exclusive). Default is 0.95.}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of various centrality measures for a given object.
}
\examples{
\dontrun{
obj <- # create your object here
  plot_centrality(obj, plot_type = "tiefighter", density_type = "outstrength", cis = 0.95)
}

}
