% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_plot.R
\name{posterior_plot}
\alias{posterior_plot}
\title{posterior_plot}
\usage{
posterior_plot(object, mat = beta, cis = c(0.8, 0.9, 0.95))
}
\arguments{
\item{object}{An object of class 'bggm'.}

\item{mat}{A matrix to use for plotting. Default is beta.}

\item{cis}{A numeric vector of credible intervals to use for plotting. Default is c(0.8, 0.9, 0.95).}
}
\description{
Plots posterior distributions of betas and partial correlations (pcors).
}
