% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{stan_fit_convert}
\alias{stan_fit_convert}
\title{Convert Stan Fit to Array of Samples}
\usage{
stan_fit_convert(stan_fit)
}
\arguments{
\item{stan_fit}{A Stan fit object obtained from either rstan or cmdstanr.}
}
\value{
A list containing three 3D arrays, each representing the posterior samples for the parameters Beta, Sigma, and Rho. Each slice of the array represents a single iteration.
}
\description{
This function converts a Stan fit object into an array of samples for each parameter (Beta, Sigma, Rho).
It supports both rstan and cmdstanr backends.
}
\examples{
\dontrun{
samples <- stan_fit_convert(stan_fit)
}

}
