% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsnet-package.R
\docType{package}
\name{tsnet-package}
\alias{tsnet-package}
\alias{tsnet}
\title{The 'tsnet' package}
\description{
Time Series Network Analysis with R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bsiepe/tsnet}
  \item Report bugs at \url{https://github.com/bsiepe/tsnet/issues}
}

}
\author{
\strong{Maintainer}: Björn S. Siepe \email{bjoernsiepe@gmail.com} (\href{https://orcid.org/0000-0002-9558-4648}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Matthias Kloft \email{kloft@uni-marburg.de} (\href{https://orcid.org/0000-0003-1845-6957}{ORCID})
}

Other contributors:
\itemize{
  \item Daniel W. Heck \email{daniel.heck@uni-marburg.de} (\href{https://orcid.org/0000-0002-6302-9252}{ORCID}) [contributor]
}

}
