% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eigen.R
\name{check_eigen}
\alias{check_eigen}
\title{Check Eigenvalues of BGGM Object}
\usage{
check_eigen(fitobj)
}
\arguments{
\item{fitobj}{A BGGM object.}
}
\value{
A list containing the eigenvalues and a verbal summary of the results.
}
\description{
This function checks the eigenvalues of Beta to assure that the model is stationary. It uses the same check as the `graphicalVAR` package.
}
