% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_gvar.R
\name{ess_gvar}
\alias{ess_gvar}
\title{Compute Effective Sample Sizes for MCMC Samples of BGGM GVAR models}
\usage{
ess_gvar(fitobj, burnin = 50)
}
\arguments{
\item{fitobj}{A `BGGM` var_estimate fit object.}

\item{burnin}{An integer indicating the number of burn-in iterations to discard. Default is 50.}
}
\value{
A list with two elements: ess_beta and ess_pcor. ess_beta contains the ESS of MCMC samples of VAR, and ess_pcor contains the ESS of MCMC samples of partial correlation coefficients.
}
\description{
This function computes the effective sample size (ESS) of MCMC samples
of temporal and contemporaneous network parameters
based on the provided `BGGM::var_estimate` fit object.
It uses the default functionality of the `coda` package to compute ESS.
}
