% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_distance_within.R
\name{post_distance_within}
\alias{post_distance_within}
\title{Calculates distances between pairs of posterior samples using the posterior samples or posterior predictive draws}
\usage{
post_distance_within(
  fitobj,
  comp,
  pred,
  draws = 1000,
  sampling_method = "random",
  indices = NULL
)
}
\arguments{
\item{fitobj}{A BGGM var_estimate fit object.}

\item{comp}{A character string indicating the type of distance between models that should be calculated. The options include: "frob" (Frobenius norm), "maxdiff" (maximum difference), or "l1" (L1 norm).}

\item{pred}{A logical indicating whether the input is posterior predictive draws (TRUE) or posterior samples (FALSE).}

\item{draws}{An integer specifying the number of random pairs of models that should be compared.}

\item{sampling_method}{Draw sequential pairs of samples from the posterior, with certain distance between them ("sequential") or randomly from two halves of the posterior ("random").
Default: "random"}

\item{indices}{A vector of indices specifying which elements of the matrices to consider when calculating distances. If NULL (default), all elements are considered. If provided, only the elements at these indices are considered. This can be useful if you want to calculate distances based on a subset of the elements in the matrices.}
}
\value{
A list of distances between the specified pairs of fitted models. The list has length equal to the specified number of random pairs. Each list element contains two distance values, one for beta coefficients and one for partial correlations.
}
\description{
This function computes distances between posterior samples of a fitted gVAR model.
Distances can be obtained either from posterior samples or posterior predictive draws.
The distance between two models can currently be calculated based on three options: Frobenius norm, maximum difference, or L1 norm.
Used within [compare_gvar()].
}
