% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_plot.R
\name{posterior_plot}
\alias{posterior_plot}
\title{posterior_plot}
\usage{
posterior_plot(fitobj, mat = "beta", cis = c(0.8, 0.9, 0.95))
}
\arguments{
\item{fitobj}{A 'var_estimate' fit object from the `BGGM` package.}

\item{mat}{A matrix to use for plotting. Possibilities include "beta" (temporal network)
and "pcor" (contemporaneous network). Default is "beta" (temporal network).}

\item{cis}{A numeric vector of credible intervals to use for plotting. Default is c(0.8, 0.9, 0.95).}
}
\description{
Plots posterior distributions of the parameters of the temporal and
contemporaneous networks of a gVAR model.
}
