% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_compare_gvar.R
\name{print.compare_gvar}
\alias{print.compare_gvar}
\title{Print method for compare_gvar objects}
\usage{
\method{print}{compare_gvar}(x, ...)
}
\arguments{
\item{x}{A compare_gvar object.}

\item{...}{Additional arguments to be passed to the print method. (currently not used)}
}
\value{
None
}
\description{
This function prints a summary of the Norm-Based Comparison Test for a [compare_gvar()] object.
}
\details{
This function prints a summary of the Norm-Based Comparison Test for a compare_gvar object.
It displays the general summary and model-specific results, including the number of significant comparisons
in the temporal and contemporaneous networks, as well as the number of reference distances that were larger
than the empirical distance for each network.
}
