% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_gvar.R
\name{stan_gvar}
\alias{stan_gvar}
\title{Fit Bayesian Generalized Vector Autoregressive (gVAR) Models with Stan}
\usage{
stan_gvar(
  data,
  beep = NULL,
  priors = NULL,
  method = "sampling",
  cov_prior = "LKJ",
  rmv_overnight = FALSE,
  iter_sampling = 500,
  iter_warmup = 500,
  n_chains = 4,
  n_cores = 4,
  center_only = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame or matrix containing the time series data.}

\item{beep}{A vector of beeps with length of `nrow(data)`. The beep indicator can be used to remove overnight effects from the last beep of a day to the first beep of the next day.}

\item{priors}{A list of prior distributions for the model parameters.}

\item{method}{A string indicating the method to use for fitting the model. Options are "sampling" or "variational".}

\item{cov_prior}{A string indicating the prior distribution to use for the covariance matrix. Options are "LKJ" or "IW" (Inverse-Wishart).}

\item{rmv_overnight}{A logical indicating whether to remove overnight effects. Default is `FALSE`.}

\item{iter_sampling}{An integer specifying the number of iterations for the sampling method. Default is 500.}

\item{iter_warmup}{An integer specifying the number of warmup iterations for the sampling method. Default is 500.}

\item{n_chains}{An integer specifying the number of chains for the sampling method. Default is 4.}

\item{n_cores}{An integer specifying the number of cores to use for parallel computation. Default is 4.}

\item{center_only}{A logical indicating whether to only center (and not scale) the data. Default is `FALSE`.}

\item{...}{Additional arguments passed to the `rstan::sampling` or `rstan::vb` function.}
}
\value{
A `stanfit` object representing the fitted model.
}
\description{
This function fits a Bayesian gVAR model to the provided data using Stan.
}
\examples{
\dontrun{
data <- matrix(rnorm(1000), ncol = 10)
fit <- stan_gvar(data, method = "sampling", cov_prior = "LKJ")
}

}
